+++ ECommerceApplication/src/main/java/com/app/entites/CartItem.java
@@ -20,7 +20,7 @@ public class CartItem {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer cartItemId;
+	private Long cartItemId;
 	
 	@ManyToOne
 	@JoinColumn(name = "cart_id")
