+++ ECommerceApplication/src/main/java/com/app/payloads/AddressDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class AddressDTO {
 
-	private Integer addressId;
+	private Long addressId;
 	private String street;
 	private String buildingName;
 	private String city;
