+++ ECommerceApplication/src/main/java/com/app/payloads/UserDTO.java
@@ -13,17 +13,14 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserDTO {
-	private Integer userId;
-
+	
+	private Long userId;
 	private String firstName;
 	private String lastName;
 	private String mobileNumber;
 	private String email;
 	private String password;
-	
 	private Set<Role> roles = new HashSet<>();
-	
 	private AddressDTO address;
-	
 	private CartDTO cart;
 }
