+++ ECommerceApplication/src/main/java/com/app/repositories/CartRepo.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import com.app.entites.Cart;
 
 @Repository
-public interface CartRepo extends JpaRepository<Cart, Integer> {
+public interface CartRepo extends JpaRepository<Cart, Long> {
 	
 	@Query("SELECT c FROM Cart c WHERE c.user.email = ?1 AND c.id = ?2")
-	Cart findCartByEmailAndCartId(String email, Integer cartId);
+	Cart findCartByEmailAndCartId(String email, Long cartId);
 }
