+++ ECommerceApplication/src/main/java/com/app/repositories/RoleRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.app.entites.Role;
 
 @Repository
-public interface RoleRepo extends JpaRepository<Role, Integer> {
+public interface RoleRepo extends JpaRepository<Role, Long> {
 
 }
