+++ ECommerceApplication/src/main/java/com/app/services/OrderService.java
@@ -5,12 +5,12 @@ import java.util.List;
 import com.app.payloads.OrderDTO;
 
 public interface OrderService {
-	OrderDTO placeOrder(String emailId, Integer cartId);
+	OrderDTO placeOrder(String emailId, Long cartId);
 	
-	OrderDTO getOrder(String emailId, Integer orderId);
+	OrderDTO getOrder(String emailId, Long orderId);
 	
 	List<OrderDTO> getAllOrders(String emailId);
 	
-	String deleteOrder(String email, Integer orderId);
+	String deleteOrder(String email, Long orderId);
 	
 }
